// src/types/promptdata.d.ts
import { Timestamp } from "firebase/firestore";

export type PromptDataType = {
  style?: string;
  mindset?: string;
  grandChallenge?: string;
  exponentialTechnology?: string;
  mtp?: string;
  moonshot?: string;
  freestyle?: string;
  downloadUrl?: string;
  prompt?: string;
  timestamp?: Timestamp;
  id?: string;
};